project/Tgbots2/requirements.txt
python-telegram-bot==20.7
httpx==0.27.0
yt-dlp==2025.1.26
ffmpeg-python==0.2.0
